rename preprocessor to precalculation
