+++ shift_detector/precalculations/text_precalculation.py
