+++ tests/precalculations/test_text_precalculation.py
@@ -3,7 +3,7 @@ import unittest
 from pandas.util.testing import assert_frame_equal
 
 from shift_detector.precalculations.store import Store
-from shift_detector.precalculations.text_preprocessor import *
+from shift_detector.precalculations.text_precalculation import *
 
 import tests.test_data as td
